add separate configuration for mail service
rename packages, clean up dependencies
