+++ src/main/java/com/planyourlifeapp/api/repository/FeedRepository.java
@@ -1,9 +1,7 @@
-package com.planyourlifeapp.api.repositorys;
+package com.planyourlifeapp.api.repository;
 
 import com.planyourlifeapp.api.models.Feed;
 import org.springframework.data.repository.PagingAndSortingRepository;
-
-import java.util.Set;
 
 public interface FeedRepository extends PagingAndSortingRepository<Feed, Long> {
     Feed getFeedById(Long id);
