+++ src/main/java/com/planyourlifeapp/api/security/AuthenticationFailureListener.java
@@ -1,10 +1,9 @@
 package com.planyourlifeapp.api.security;
 
-import com.planyourlifeapp.api.services.LoginAttemptService;
+import com.planyourlifeapp.api.service.LoginAttemptService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationListener;
 import org.springframework.security.authentication.event.AuthenticationFailureBadCredentialsEvent;
-import org.springframework.security.web.authentication.WebAuthenticationDetails;
 import org.springframework.stereotype.Component;
 
 @Component
