+++ src/main/java/com/planyourlifeapp/api/security/AuthenticationSuccessEventListener.java
@@ -1,6 +1,6 @@
 package com.planyourlifeapp.api.security;
 
-import com.planyourlifeapp.api.services.LoginAttemptService;
+import com.planyourlifeapp.api.service.LoginAttemptService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationListener;
 import org.springframework.security.authentication.event.AuthenticationSuccessEvent;
